Fix & Move tests from test_projects to test_utils
