+++ tests/projects/conftest.py
@@ -1,0 +1,35 @@
+import os
+
+import git
+from distutils import dir_util
+import pytest
+
+import mlflow
+from mlflow.utils.file_utils import path_to_local_sqlite_uri
+
+from tests.projects.utils import TEST_PROJECT_DIR
+
+
+@pytest.fixture
+def local_git_repo(tmpdir):
+    local_git = tmpdir.join('git_repo').strpath
+    repo = git.Repo.init(local_git)
+    dir_util.copy_tree(src=TEST_PROJECT_DIR, dst=local_git)
+    dir_util.copy_tree(src=os.path.dirname(TEST_PROJECT_DIR), dst=local_git)
+    repo.git.add(A=True)
+    repo.index.commit("test")
+    yield os.path.abspath(local_git)
+
+
+@pytest.fixture
+def local_git_repo_uri(local_git_repo):
+    return "file://%s" % local_git_repo
+
+
+@pytest.fixture()
+def tracking_uri_mock(tmpdir):
+    try:
+        mlflow.set_tracking_uri(path_to_local_sqlite_uri(os.path.join(tmpdir.strpath, 'mlruns')))
+        yield tmpdir
+    finally:
+        mlflow.set_tracking_uri(None)
